# Conflicts:
#	src/main/java/uk/ac/ebi/pride/spectracluster/cluster/SpectralQualityHolder.java
#	src/main/java/uk/ac/ebi/pride/spectracluster/io/BinaryClusterParser.java
#	src/main/java/uk/ac/ebi/pride/spectracluster/io/DotClusterClusterAppender.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/cluster/ClusterComparatorTests.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/cluster/IncrementalClusteringEngineTests.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/consensus/ConsensusSpectrumTests.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/consensus/JohannesConsensusSpectrum.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/engine/GreedySpectralClusteringEngineTest.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/io/GreedyParserUtilitiesTests.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/io/RetentionTimeTest.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/quality/SignalToNoiseQualityCheckerTests.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/util/JMZTabUtilities.java
#	src/test/java/uk/ac/ebi/pride/spectracluster/util/predicate/spectrum_comparison/ShareMajorPeaksPredicateTest.java
#	src/test/java/uk/ac/ebi/pride/tools/pride_spectra_clustering/impl/Adapters.java

+++ src/main/java/uk/ac/ebi/pride/spectracluster/similarity/FrankEtAlDotProduct.java
@@ -151,14 +151,13 @@ public class FrankEtAlDotProduct implements ISimilarityChecker {
 
             highestPeaksSpectrum1 = spectrum1.getHighestNPeaks(numberCompared);
             highestPeaksSpectrum2 = spectrum2.getHighestNPeaks(numberCompared);
-        }
-        else {
+        } else {
             // don't use peak filtering
             highestPeaksSpectrum1 = spectrum1;
             highestPeaksSpectrum2 = spectrum2;
         }
 
-        IPeakMatches peakMatches = PeakMatchesUtilities.getSharedPeaksAsMatches(highestPeaksSpectrum1, highestPeaksSpectrum2, (float) this.fragmentIonTolerance);
+        IPeakMatches peakMatches = PeakMatchesUtilities.getSharedPeaksAsMatches(highestPeaksSpectrum1, highestPeaksSpectrum2, this.fragmentIonTolerance);
 
         return assessSimilarity(peakMatches);
     }
